+++ TombEngine/Objects/TR3/Trap/TurningBlade.h
