+++ TombEngine/TombEngine.vcxproj
@@ -627,7 +627,7 @@ if not exist "%ScriptsDir%\Strings.lua" xcopy /Y "$(SolutionDir)Scripts\Strings.
     <ClInclude Include="Objects\TR3\Object\Corpse.h" />
     <ClInclude Include="Objects\TR3\Trap\ElectricCleaner.h" />
     <ClInclude Include="Objects\TR3\Trap\train.h" />
-    <ClInclude Include="Objects\TR3\Trap\turning_blade.h" />
+    <ClInclude Include="Objects\TR3\Trap\TurningBlade.h" />
     <ClInclude Include="Objects\TR3\Trap\WallMountedBlade.h" />
     <ClInclude Include="Objects\TR3\Vehicles\big_gun.h" />
     <ClInclude Include="Objects\TR3\Vehicles\big_gun_info.h" />
@@ -1180,7 +1180,7 @@ if not exist "%ScriptsDir%\Strings.lua" xcopy /Y "$(SolutionDir)Scripts\Strings.
     <ClCompile Include="Objects\TR3\tr3_objects.cpp" />
     <ClCompile Include="Objects\TR3\Trap\ElectricCleaner.cpp" />
     <ClCompile Include="Objects\TR3\Trap\train.cpp" />
-    <ClCompile Include="Objects\TR3\Trap\turning_blade.cpp" />
+    <ClCompile Include="Objects\TR3\Trap\TurningBlade.cpp" />
     <ClCompile Include="Objects\TR3\Trap\WallMountedBlade.cpp" />
     <ClCompile Include="Objects\TR3\Vehicles\big_gun.cpp" />
     <ClCompile Include="Objects\TR3\Vehicles\kayak.cpp" />
